% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfish.R
\name{pfish}
\alias{pfish}
\title{Fish dynamics (Biomass)}
\usage{
pfish(x)
}
\arguments{
\item{Please}{use the estimates of 'pphy'}
}
\description{
This function allows you to express the fish dynamics using GOCI
}
\examples{
pfish()
}
\keyword{Fish}
