% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pphy.R
\name{pphy}
\alias{pphy}
\title{Phytoplankton dynamics (Biomass)}
\usage{
pphy(x)
}
\arguments{
\item{Please}{use the Chl-a pixel data offered by GOCI}
}
\description{
This function allows you to express the phytoplankton dynamics using GOCI
}
\examples{
pphy()
}
\keyword{Phytoplankton}
