% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pzoo.R
\name{pzoo}
\alias{pzoo}
\title{Zooplankton dynamics (Biomass)}
\usage{
pzoo(x)
}
\arguments{
\item{Please}{use the estimates of 'pphy'}
}
\description{
This function allows you to express the zooplankton dynamics using GOCI
}
\examples{
pzoo()
}
\keyword{Zooplankton}
